bug(#7974): give three more fixtures classes of their own
The three recursion fixtures this branch used to move are already out,
lifted by dd1f9396f7 in #8550. `Rnd`, `WithVoidPhi` and `Counter` are
the next three that no other test reads, so they move next to
`EndlessRecursion` and friends, and `PhDefaultTest` drops from 901 to
820 lines.

`Foo` and `Kid` are left for a follow-up, since `Foo` makes a `Kid` and
the pair has to travel together, and so is `Int`, which cannot keep its
name at the top level next to `org.eolang.Int`. Both would push this
over the 200-line pr-size limit anyway.
